Fix representation problem of doubles
When a double is entered. Not using the string method
ends with the number being not stored with the proper decimal
ie. 2.3 where estores as 2.2999999200002 or similar. This then
affect to the java manipulation of this parameter.

The changes keep with the same resolution and possible decimals
but when it stores 2.3 konw also return 2.3 where the value is
accessed.